+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 * Upgraded to `polymer-analyzer@3.0.0-pre.17` and `polymer-bundler@4.0.0-pre.2`.
 * Converted from `PackageUrlResolver` to `FsUrlResolver` as more appropriate to application build context.
+* Ignore `not-loadable` warnings, as polymer-build should not load remote
+  resources.
 <!-- Add new, unreleased changes here. -->
 
 ## [2.5.0] - 2018-03-21